Use test directory structure in tracpro.contacts
